scripts/requirements.txt
2captcha-python==1.5.0
absl-py==1.4.0
aenum==2.2.6
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.1
alembic==1.14.0
anaconda==0.0.1.1
annotated-types==0.6.0
anthropic==0.49.0
anyio==4.7.0
apache-airflow==2.10.4
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apispec==6.8.0
appnope==0.1.3
argcomplete==3.5.2
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==24.2.0
autopep8==2.0.0
babel==2.16.0
backcall==0.2.0
BAScraper==0.0.3
beautifulsoup4==4.12.3
binaryornot==0.4.4
biplist==1.0.3
blinker==1.9.0
blis==1.2.0
cachelib==0.9.0
cachetools==5.2.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
comm==0.1.3
commonmark==0.9.1
confection==0.1.5
ConfigUpdater==3.2
connexion==2.14.2
cookiecutter==2.6.0
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
cymem==2.0.11
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
Django==4.0.1
dm-tree==0.1.8
dnspython==2.7.0
dotenv==0.9.9
dotenv-python==0.0.1
email_validator==2.2.0
et-xmlfile==1.1.0
etils==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
fastapi==0.115.10
fastapi-cli==0.0.7
fbchat==1.9.7
filelock==3.12.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
functions-framework==3.2.0
gensim==4.3.3
GeoAlchemy2==0.12.5
geojson==2.5.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-re2==1.1.20240702
google-resumable-media==2.4.0
googleapis-common-protos==1.66.0
gotrue==2.11.2
greenlet==2.0.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.21.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.15.1
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.3.0
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.10.0
limits==3.14.1
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==4.9.1
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
murmurhash==1.0.12
nest-asyncio==1.5.6
networkx==3.4.2
numpy==1.24.2
oauthlib==3.2.2
openai==1.25.1
openpyxl==3.1.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
outcome==1.2.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
parso==0.8.3
pathspec==0.12.1
pdf2image==1.17.0
pendulum==3.0.0
pexpect==4.8.0
pg8000==1.29.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.0
pluggy==1.5.0
postgrest==0.19.3
preshed==3.0.9
prison==0.2.1
promise==2.3
prompt-toolkit==3.0.38
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.0.2
realtime==2.2.0
referencing==0.35.1
regex==2022.10.31
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.1
scipy==1.10.1
scramp==1.4.4
segeval==2.0.11
selenium==4.7.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.2
staffspy==0.2.25
starlette==0.46.0
storage3==0.11.1
StrEnum==0.4.15
stweet==2.1.1
supabase==2.12.0
supafunc==0.9.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorflow-datasets==4.8.3
tensorflow-metadata==1.12.0
termcolor==2.5.0
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.1.0
time-machine==2.16.0
tldextract==5.1.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tor-python-easy==0.1.7
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchdog==2.1.9
watchfiles==1.0.4
wcwidth==0.2.6
weasel==0.4.1
webdriver-manager==4.0.1
websockets==13.1
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.2.0
yarl==1.18.3
zipp==3.21.0
